+++ test/conftest.py
@@ -1,16 +1,21 @@
 import uuid
+from pathlib import Path
 
 import geopandas as gpd
 import pytest
 import shapely
 from climatoology.base.baseoperator import AoiProperties
 from climatoology.base.computation import ComputationScope
+from dotenv import load_dotenv
 from ohsome_py2.client import OhsomeClient
 from shapely.geometry.polygon import Polygon
 
 from land_consumption.components.landuse_category_mappings import LandObjectCategory
 from land_consumption.core.input import ComputeInput
 from land_consumption.core.operator_worker import LandConsumption
+
+load_dotenv()  # To load the `OHSOME_BASE_URL` environment variable, for recording new cassettes
+pytest_plugins = ('ohsome_py2.test.fixtures',)
 
 
 @pytest.fixture
@@ -47,13 +52,13 @@ def compute_resources():
 
 
 @pytest.fixture
-def default_operator(default_ohsome_client_v1):
-    return LandConsumption(ohsome_client=default_ohsome_client_v1)
+def default_operator(parametrized_ohsome_client):
+    return LandConsumption(ohsome_client=parametrized_ohsome_client)
 
 
-@pytest.fixture
-def default_ohsome_client_v1():
-    return OhsomeClient(user_agent='Land-Consumption Test', v2=False)
+@pytest.fixture(params=[True, False])
+def parametrized_ohsome_client(request):
+    return OhsomeClient(user_agent='Land-Consumption Test', v2=request.param)
 
 
 @pytest.fixture(scope='module')
@@ -94,8 +99,17 @@ def categories_gdf():
 
 
 @pytest.fixture(scope='module')
-def vcr_config():
+def vcr_cassette_dir(request):
+    file_name = Path(request.module.__file__).stem
+    cassette_dir = str(Path(__file__).parent / 'resources' / 'vcr_cassettes' / file_name)
+    return cassette_dir
+
+
+@pytest.fixture(scope='module')
+def vcr_config(vcr_config_ohsomepy2):
     return {
+        **vcr_config_ohsomepy2,
         'filter_headers': ['authorization'],
-        'cassette_library_dir': 'test/resources/vcr_cassettes',
+        # Match on the default parts plus the request body
+        'match_on': ['method', 'scheme', 'host', 'port', 'path', 'query', 'body'],
     }